requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
atomicwrites==1.4.0
attrs==20.2.0
colorama==0.4.4
importlib-metadata==2.0.0
iniconfig==1.1.1
more-itertools==8.5.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==5.1.1
pytest-rerunfailures==9.0
python-dateutil==2.8.1
selenium==3.14.0
six==1.15.0
text-unidecode==1.3
toml==0.10.2
urllib3==1.25.10
wcwidth==0.2.5
zipp==3.3.1
